rooms: access control
https://trello.com/c/jYgMC6J9/1441-hausbewohnerverwaltung

+++ app/controllers/api/v1/rooms_controller.rb
@@ -15,21 +15,22 @@ class Api::V1::RoomsController < Api::V1::BaseController
   api :POST, '/api/v1/corporations/ID/rooms'
 
   def create
-    authorize! :update, corporation
+    authorize! :update_accommodations, corporation
 
     new_room = corporation.create_room room_params
     render json: new_room, status: :ok
   end
 
   def update
-    authorize! :update, room
+    authorize! :update_accommodations, room.corporation
 
     room.update! room_params
     render json: room, status: :ok
   end
 
   def destroy
-    authorize! :destroy, room
+    authorize! :update_accommodations, room.corporation
+    raise "Cannot remove room if it has current or past occupants" if room.memberships.with_past.any?
 
     room.destroy!
     render json: {}, status: :ok
